+++ src/main/java/roomescape/controller/dto/CreateUserReservationRequest.java
@@ -1,6 +1,6 @@
 package roomescape.controller.dto;
 
-import roomescape.global.exception.RoomescapeException;
+import roomescape.system.exception.RoomescapeException;
 
 public record CreateUserReservationRequest(String date, Long themeId, Long timeId) {
