+++ src/main/java/roomescape/service/LoginService.java
@@ -4,9 +4,9 @@ import org.springframework.stereotype.Service;
 import roomescape.domain.member.Email;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.Password;
-import roomescape.global.auth.JwtManager;
-import roomescape.global.exception.AuthorizationException;
 import roomescape.repository.MemberRepository;
+import roomescape.system.auth.JwtManager;
+import roomescape.system.exception.AuthorizationException;
 
 @Service
 public class LoginService {
@@ -20,7 +20,8 @@ public class LoginService {
     }
 
     public String login(String email, String password) {
-        Member member = memberRepository.findByEmailAndPassword(new Email(email), new Password(password))
+        Member member = memberRepository.findByEmailAndPassword(new Email(email),
+                new Password(password))
             .orElseThrow(() -> new AuthorizationException("아이디 혹은 패스워드가 일치하지 않습니다."));
 
         return jwtManager.createToken(member);
