+++ src/main/java/roomescape/system/argumentresolver/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package roomescape.global.argumentresolver;
+package roomescape.system.argumentresolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
