+++ src/main/java/roomescape/system/auth/CheckUserInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.global.auth;
+package roomescape.system.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
